@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.Text;
 
@@ -8,18 +9,32 @@ namespace YogaMo.Model.Requests
     public class InstructorsInsertRequest
     {
         [Required]
+        [StringLength(50)]
         public string FirstName { get; set; }
+        [Required]
+        [StringLength(50)]
         public string LastName { get; set; }
-        [EmailAddress]
+        [Required]
+        [DataType(DataType.EmailAddress)]
+        [StringLength(50)]
+        [RegularExpression(@"(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|""(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*"")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])")]
         public string Email { get; set; }
+        [Required]
+        [StringLength(25)]
+        [DataType(DataType.PhoneNumber)]
         public string Phone { get; set; }
+        [Required]
+        [StringLength(15, MinimumLength = 4)]
         public string Username { get; set; }
-        public bool? Status { get; set; }
         public string Title { get; set; }
-        //   public int? CityId { get; set; }
 
-        // public City City { get; set; }
+        [Required]
+        [StringLength(10, MinimumLength = 4)]
+        [DataType(DataType.Password)]
         public string Password { get; set; }
+        [Required]
+        [DataType(DataType.Password)]
         public string PasswordConfirmation { get; set; }
+        public byte[] ProfilePicture { get; set; }
     }
 }
