@@ -10,8 +10,9 @@ namespace YogaMo.WebAPI.Database
         public int? OrderId { get; set; }
         public int Quantity { get; set; }
         public decimal Price { get; set; }
+        public string Size { get; set; }
 
-        public Order Order { get; set; }
-        public Product Product { get; set; }
+        public virtual Order Order { get; set; }
+        public virtual Product Product { get; set; }
     }
 }
