@@ -1,18 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 
 namespace YogaMo.WebAPI.Database
 {
-    public partial class OrderStatus
+    public enum OrderStatus
     {
-        public OrderStatus()
-        {
-            Order = new HashSet<Order>();
-        }
-
-        public int OrderStatusId { get; set; }
-        public string Status { get; set; }
-
-        public ICollection<Order> Order { get; set; }
+        New,
+        Processing,
+        Completed,
+        Canceled
     }
 }
