@@ -7,7 +7,7 @@ namespace YogaMo.WebAPI.Exceptions
 {
     public class UserException : Exception
     {
-        public UserException(string message): base(message)
+        public UserException(string message) : base(message)
         {
 
         }
