@@ -20,13 +20,15 @@ namespace YogaMo.WebAPI.Filters
             }
             else
             {
-                context.ModelState.AddModelError("ERROR", "Greška na serveru");
+                context.ModelState.AddModelError("ERROR", context.Exception.GetExceptionMessages());
                 context.HttpContext.Response.StatusCode = (int)HttpStatusCode.InternalServerError;
             }
 
-            var list = context.ModelState.Where(x => x.Value.Errors.Count > 0).ToDictionary(x => x.Key, y => y.Value.Errors.Select(z => z.ErrorMessage));
+            /*  var list = context.ModelState.Where(x => x.Value.Errors.Count > 0).ToDictionary(x => x.Key, y => y.Value.Errors.Select(z => z.ErrorMessage));*/
 
-            context.Result = new JsonResult(list);
+            //  context.Result = new JsonResult(list);
+
+            context.Result = new JsonResult(context.ModelState);
         }
     }
 }
